feat: adds app subcommands, refactors interface into noun-verb structure
