+++ f76/cli.py
@@ -1,6 +1,7 @@
 import os, pathlib, typer
 from rich.console import Console
-from rich.table import Table
+
+from .cli_utils.cli_presentation import make_pipboy_table
 from .scripts.scrape.junk_items_table import main as scrape_junk_items
 from .scripts.scrape.regions_and_locations import main as scrape_regions_and_locations
 from .scripts.scrape.junk_locations import scrape_item_locations_by_name
@@ -9,26 +10,18 @@ from .scripts.db_utils import fetch_all
 from rich import box
 
 app = typer.Typer(help="Fallout 76 Personal Data Assistant")
+# Create sub app for registering our different items that can be acted on
+junk_app = typer.Typer(help="Explore & analyze junk items")
+scrap_app = typer.Typer(help="Explore & analyze scrap/components")
+location_app = typer.Typer(help="Explore & analyze locations")
+region_app = typer.Typer(help="Explore and analyze regions of Appalachia")
+
+app.add_typer(junk_app, name="junk")
+app.add_typer(scrap_app, name="scrap")
+app.add_typer(location_app, name="location")
+app.add_typer(region_app, name="region")
+
 console = Console()
-
-PRIMARY_GREEN = "#03e903"
-SECONDARY_GREEN = "#03AF03"
-
-def make_pipboy_table(title: str, width: int = 60) -> Table:
-    """
-    Creates table styled like a Fallout Pip-Boy.
-    Keeps consistent headers, colors, and spacing across commands.
-    """
-    return Table(
-        title=f"[{PRIMARY_GREEN}]{title}[/{PRIMARY_GREEN}]",
-        expand=False,
-        width=width,
-        show_lines=True,
-        padding=(0, 1),
-        header_style=f"bold {PRIMARY_GREEN}",
-        border_style=PRIMARY_GREEN,
-        row_styles=[PRIMARY_GREEN, SECONDARY_GREEN] # alternating row background
-    )
 
 def default_data_dir() -> pathlib.Path:
     base = pathlib.Path.home() / ".local" / "share" / "f76"  # fine on mac/Linux
@@ -53,7 +46,8 @@ def resolve_db_path(db_opt: str | None = None) -> pathlib.Path:
     # user data dir fallback
     return default_data_dir() / "fallout.sqlite"
 
-@app.command("scrap")
+# Register this action under the new junk_app
+@junk_app.command("scrap")
 def scrap(item: str, db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
     """
     Look up what components a Junk Item will scrap into (example: `f76 scrap 'Giddyup Buttercup'`)
@@ -77,98 +71,8 @@ def scrap(item: str, db: str | None = typer.Option(None, help="Path to fallout.s
         t.add_row(comp, str(qty))
     console.print(t)
 
-@app.command("sources")
-def sources(component: str, db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
-    """
-    Look up what Junk Items are a source of a given component (example: `f76 sources 'Lead'`)
-    """
-    q = """
-    SELECT i.name, s.quantity
-    FROM component c
-    JOIN item_scraps s ON s.component_id = c.id
-    JOIN item        i ON i.id = s.item_id
-    WHERE c.name = ? COLLATE NOCASE
-    ORDER BY s.quantity DESC, i.name;
-    """
-    db_path = resolve_db_path(db)
-    rows, _ = fetch_all(db_path, q, (component,)) 
-    if not rows:
-        console.print(f"[bold]No items found for component:[/bold] {component} (DB: {db_path})")
-        raise typer.Exit(1)
-    t = make_pipboy_table(f'Items that yield "{component}":')
-    t.add_column("Item"); t.add_column("Qty", justify="right")
-    for item_name, qty in rows:
-        t.add_row(item_name, str(qty))
-    console.print(t)
-
-@app.command("whereis")
-def region_for(location: str,db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
-    """
-    Look up what region a location exists in. (example: `f76 whereis 'Wade Airport'`)
-    """
-    q = """
-    SELECT r.name
-    FROM region r
-    JOIN location l ON l.region_id = r.id
-    WHERE l.name = ? COLLATE NOCASE
-    """
-    db_path = resolve_db_path(db)
-    rows, _ = fetch_all(db_path, q, (location,))
-    if not rows:
-        console.print(f"[bold]No region found for location:[/bold] {location.title()} (DB: {db_path})")
-        raise typer.Exit(1)
-    t = make_pipboy_table(f'{location.title()} is located in:')
-    t.add_column("Region");
-    # fetchall returns a list of tuples - even for one col
-    for (region,) in rows:
-        t.add_row(region)
-    console.print(t)
-
-@app.command("places")
-def locations_in(region: str,db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
-    """
-    Look up what locations are in a region of the map (example: `f76 places 'Cranberry Bog'`)
-    """
-    q = """
-    SELECT l.name
-    FROM location l
-    JOIN region r ON l.region_id = r.id
-    WHERE r.name = ? COLLATE NOCASE
-    ORDER BY l.name
-    """
-    db_path = resolve_db_path(db)
-    rows, _ = fetch_all(db_path, q, (region,))
-    if not rows:
-        console.print(f"[bold]No locations found for region:[/bold] {region.title()}")
-        raise typer.Exit(1)
-    t = make_pipboy_table(f'{region.title()} is home to the following locations:')
-    t.add_column("Locations");
-    for (location_name,) in rows:
-        t.add_row(location_name)
-    console.print(t)
-
-@app.command("regions")
-def locations_in(db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
-    """
-    List all the regions of the map (example: `f76 regions`)
-    """
-    q = """
-    SELECT r.name
-    FROM region r
-    ORDER BY r.name
-    """
-    db_path = resolve_db_path(db)
-    rows, _ = fetch_all(db_path, q)
-    if not rows:
-        console.print(f"[bold]No regions found.[/bold] Have you ran `f76 init`?")
-        raise typer.Exit(1)
-    t = make_pipboy_table(f'You will find the following Regions in Appalachia:')
-    t.add_column("Regions");
-    for (region_name,) in rows:
-        t.add_row(region_name)
-    console.print(t)
-
-@app.command("where")
+# Junk Command
+@junk_app.command("find")
 def where(item: str, db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
     """
     List locations where a junk item can be found (example `f76 where soap`)
@@ -203,6 +107,101 @@ def where(item: str, db: str | None = typer.Option(None, help="Path to fallout.s
         t.add_row(loc_name, str(qty) if qty is not None else "-", desc)
     console.print(t)
         
+
+# Went with "scrap" because the community recognizes components as "scrap"
+@scrap_app.command("sources")
+def sources(component: str, db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
+    """
+    Look up what Junk Items are a source of a given component (example: `f76 sources 'Lead'`)
+    """
+    q = """
+    SELECT i.name, s.quantity
+    FROM component c
+    JOIN item_scraps s ON s.component_id = c.id
+    JOIN item        i ON i.id = s.item_id
+    WHERE c.name = ? COLLATE NOCASE
+    ORDER BY s.quantity DESC, i.name;
+    """
+    db_path = resolve_db_path(db)
+    rows, _ = fetch_all(db_path, q, (component,)) 
+    if not rows:
+        console.print(f"[bold]No items found for component:[/bold] {component} (DB: {db_path})")
+        raise typer.Exit(1)
+    t = make_pipboy_table(f'Items that yield "{component}":')
+    t.add_column("Item"); t.add_column("Qty", justify="right")
+    for item_name, qty in rows:
+        t.add_row(item_name, str(qty))
+    console.print(t)
+
+@location_app.command("find")
+def region_for(location: str,db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
+    """
+    Look up what region a location exists in. (example: `f76 whereis 'Wade Airport'`)
+    """
+    q = """
+    SELECT r.name
+    FROM region r
+    JOIN location l ON l.region_id = r.id
+    WHERE l.name = ? COLLATE NOCASE
+    """
+    db_path = resolve_db_path(db)
+    rows, _ = fetch_all(db_path, q, (location,))
+    if not rows:
+        console.print(f"[bold]No region found for location:[/bold] {location.title()} (DB: {db_path})")
+        raise typer.Exit(1)
+    t = make_pipboy_table(f'{location.title()} is located in:')
+    t.add_column("Region");
+    # fetchall returns a list of tuples - even for one col
+    for (region,) in rows:
+        t.add_row(region)
+    console.print(t)
+
+# Conceptually different from the ones above - "find doesn't fit in as well here"
+@region_app.command("locations")
+def locations_in(region: str,db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
+    """
+    Look up what locations are in a region of the map (example: `f76 places 'Cranberry Bog'`)
+    """
+    q = """
+    SELECT l.name
+    FROM location l
+    JOIN region r ON l.region_id = r.id
+    WHERE r.name = ? COLLATE NOCASE
+    ORDER BY l.name
+    """
+    db_path = resolve_db_path(db)
+    rows, _ = fetch_all(db_path, q, (region,))
+    if not rows:
+        console.print(f"[bold]No locations found for region:[/bold] {region.title()}")
+        raise typer.Exit(1)
+    t = make_pipboy_table(f'{region.title()} is home to the following locations:')
+    t.add_column("Locations");
+    for (location_name,) in rows:
+        t.add_row(location_name)
+    console.print(t)
+
+@region_app.command("list")
+def locations_in(db: str | None = typer.Option(None, help="Path to fallout.sqlite")):
+    """
+    List all the regions of the map (example: `f76 regions`)
+    """
+    q = """
+    SELECT r.name
+    FROM region r
+    ORDER BY r.name
+    """
+    db_path = resolve_db_path(db)
+    rows, _ = fetch_all(db_path, q)
+    if not rows:
+        console.print(f"[bold]No regions found.[/bold] Have you ran `f76 init`?")
+        raise typer.Exit(1)
+    t = make_pipboy_table(f'You will find the following Regions in Appalachia:')
+    t.add_column("Regions");
+    for (region_name,) in rows:
+        t.add_row(region_name)
+    console.print(t)
+
+
 # TODO: 
 # @app.command("enemy")
